done changes in feature/GAValidateNegative branch
