+++ src/main/java/com/training/services/ga/validate/Error.java
@@ -1,4 +1,4 @@
-package com.training.services.ga;
+package com.training.services.ga.validate;
 
 public class Error {
     String errorCode;
