+++ src/test/java/com/training/services/tests/GAAccountTest.java
@@ -1,7 +1,7 @@
-package services.tests;
+package com.training.services.tests;
 
 import com.training.base.BaseTest;
-import com.training.services.ga.GAValidationResponse;
+import com.training.services.ga.validate.Response;
 import com.training.utilities.RestEngine;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -9,39 +9,8 @@ import org.junit.jupiter.api.Test;
 import java.util.HashMap;
 import java.util.Map;
 
-public class GuestAccountValidateTest extends BaseTest {
-
+public class GAAccountTest extends BaseTest {
     Map<String, Object> headerMap = new HashMap();
-
-    /**
-     * Positive validation, putting all valid details and checking response
-     */
-    @Test
-    public void testGuestAccountValidation() {
-        headerMap.put(map.get("AppKeyHeader"), map.get("AppKeyValue"));
-        headerMap.put(map.get("ContentTypeHeader"), map.get("ContentTypeValue"));
-
-        RestEngine restEngine = new RestEngine();
-
-        GAValidationResponse gaValidationResponse =
-                restEngine.getResponsePost(map.get("base_url") + "/validation"
-                        , headerMap
-                        , "{\"email\": \"testPranav@api.com\"}")
-                        .as(GAValidationResponse.class);
-
-        Assertions.assertEquals(gaValidationResponse.getStatus(), 200, "Json response is not 200");
-
-        softAssert.assertThat(gaValidationResponse.getErrors().toString().length() == 0)
-                .as("Errors array is not null");
-
-        softAssert.assertThat(gaValidationResponse.getPayload().getAccountStatus())
-                .as("Inside payload JSON, accountStatus is not equals to EXISTS")
-                .isEqualTo("EXISTS");
-
-        softAssert.assertThat(gaValidationResponse.getPayload().getIsUid())
-                .as("Inside payload JSON, isUid should be true")
-                .isEqualTo(true);
-    }
 
     /**
      * Putting headers AppKey wrong and validating response
@@ -51,13 +20,11 @@ public class GuestAccountValidateTest extends BaseTest {
         headerMap.put(map.get("AppKeyHeader"), map.get("WrongAppKeyValue"));
         headerMap.put(map.get("ContentTypeHeader"), map.get("ContentTypeValue"));
 
-        RestEngine restEngine = new RestEngine();
-
-        GAValidationResponse gaValidationNegativeResponse =
-                restEngine.getResponsePost(map.get("base_url") + "/validation"
+        Response gaValidationNegativeResponse =
+                new RestEngine().getResponsePost(map.get("base_url") + "/validation"
                         , headerMap
                         , "{\"email\": \"testPranav@api.com\"}")
-                        .as(GAValidationResponse.class);
+                        .as(Response.class);
 
         Assertions.assertEquals(gaValidationNegativeResponse.getStatus(), 401
                 , "Json response is not 401");
@@ -72,7 +39,7 @@ public class GuestAccountValidateTest extends BaseTest {
                 .isEqualTo("The API key header is required and the API key provided should be valid.");
 
 
-        softAssert.assertThat(gaValidationNegativeResponse.getErrors().getInternalMessage())
+        softAssert.assertThat(gaValidationNegativeResponse.getErrors().get(0).getInternalMessage())
                 .as("Inside errors JSON, message is not equals to" +
                         "The API key header is required and should be valid.")
                 .isEqualTo("The API key header is required and should be valid.");
@@ -86,13 +53,11 @@ public class GuestAccountValidateTest extends BaseTest {
         headerMap.put(map.get("AppKeyHeader"), map.get("AppKeyValue"));
         headerMap.put(map.get("ContentTypeHeader"), map.get("ContentTypeValue"));
 
-        RestEngine restEngine = new RestEngine();
-
-        GAValidationResponse gaValidationResponse =
-                restEngine.getResponsePost(map.get("base_url") + "/validation"
+        Response gaValidationResponse =
+                new RestEngine().getResponsePost(map.get("base_url") + "/validation"
                         , headerMap
                         , "{\"email\": \"assignment70test@api.com\"}")
-                        .as(GAValidationResponse.class);
+                        .as(Response.class);
 
         Assertions.assertEquals(gaValidationResponse.getStatus(), 200, "Json response is not 200");
 
@@ -113,23 +78,21 @@ public class GuestAccountValidateTest extends BaseTest {
         headerMap.put(map.get("AppKeyHeader"), map.get("AppKeyValue"));
         headerMap.put(map.get("ContentTypeHeader"), map.get("ContentTypeValue"));
 
-        RestEngine restEngine = new RestEngine();
-
-        GAValidationResponse gaValidationNegativeResponse =
-                restEngine.getResponsePost(map.get("base_url") + "/validation"
+        Response gaValidationNegativeResponse =
+                new RestEngine().getResponsePost(map.get("base_url") + "/validation"
                         , headerMap
                         , "{\"email\": \"testPranav@@api.com\"}")
-                        .as(GAValidationResponse.class);
+                        .as(Response.class);
 
         Assertions.assertEquals(gaValidationNegativeResponse.getStatus(), 422
                 , "Json response is not 422");
 
-        softAssert.assertThat(gaValidationNegativeResponse.getErrors().getDeveloperMessage())
+        softAssert.assertThat(gaValidationNegativeResponse.getErrors().get(0).getDeveloperMessage())
                 .as("Inside errors JSON, developer message is not equal to," +
                         " The request body did not pass input validation.")
                 .isEqualTo("The request body did not pass input validation.");
 
-        softAssert.assertThat(gaValidationNegativeResponse.getErrors().getErrorCode())
+        softAssert.assertThat(gaValidationNegativeResponse.getErrors().get(0).getErrorCode())
                 .as("Inside error JSON, error code is not equals to GA-0103")
                 .isEqualTo("GA-0103");
     }
